AuthorizeNoteCommandAsync unit tests
- Introduced IUserAccessCalculation interface and UserAccessCalculation class to encapsulate user access logic.
- Updated AuthorizationEngine to utilize UserAccessCalculation for user access authorization.
- Removed AuthorizeUserAccessAsync method from IAuthorizationEngine interface.
- Adjusted dependency injection in Startup.cs to include UserAccessCalculation.
- Modified MembershipManager and RecordsManager to use UserAccessCalculation instead of AuthorizationEngine for user access checks.
- Added unit tests for AuthorizeNoteCommandAsync to validate new user access logic.
- Refactored existing tests to accommodate changes in authorization logic.
